+++ format/perfect_ref.rt
@@ -7,4 +7,4 @@ pl 0,1,0 0,-1,0 255,255,255
 pl 1,0,0 -1,0,0 255,0,255
 pl -1,0,0 1,0,0 255,255,0
 pl 0,0,5 0,0,-1 0,255,255
-sp -0.25,-0.5,3 0.5 255,0,0
+sp -0.25,-0.5,3 0.5 255,0,0 PERFECT
